started fetch decode execute thing
right now if you run it with -rom whatever.lol it will step through and
printf the op codes it has. these aren’t implemented, its just logging
them.
if one isn’t in the massive switch, it’ll output a “?”.
later on, we should put an assert in the “?” so we can implement it
then and there.